requirements.txt
altgraph==0.17.4
colorama==0.4.6
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pywin32-ctypes==0.2.3
tqdm==4.67.1
